+++ tests/e2e/multicluster/verify-install/defaultresource_test.go
@@ -4,11 +4,9 @@
 package defaultresource_test
 
 import (
-	"encoding/json"
 	"fmt"
 	"github.com/verrazzano/verrazzano/pkg/test/framework"
 	"os"
-	"regexp"
 	"time"
 
 	. "github.com/onsi/ginkgo/v2"
@@ -20,10 +18,8 @@ import (
 )
 
 const (
-	waitTimeout           = 5 * time.Minute
-	pollingInterval       = 10 * time.Second
-	cattleSystemNamespace = "cattle-system"
-	searchTimeWindow      = "1h"
+	waitTimeout     = 5 * time.Minute
+	pollingInterval = 10 * time.Second
 )
 
 var expectedPodsKubeSystem = []string{
@@ -65,79 +61,7 @@ var _ = t.Describe("Multi Cluster Install Validation", Label("f:platform-lcm.ins
 				}, waitTimeout, pollingInterval).Should(BeTrue())
 			})
 		})
-
-		t.It("Rancher log records do not contain any websocket bad handshake messages", func() {
-			// GIVEN existing system logs
-			// WHEN the Elasticsearch index for the cattle-system namespace is retrieved
-			// THEN is has a limited number of bad socket messages
-			indexName, err := pkg.GetOpenSearchSystemIndex(cattleSystemNamespace)
-			Expect(err).To(BeNil())
-			Eventually(func() bool {
-				return pkg.LogIndexFound(indexName)
-			}, waitTimeout, pollingInterval).Should(BeTrue(), "Expected to find Elasticsearch index cattle-system")
-
-			Expect(getNumBadSocketMessages()).To(BeNumerically("<", 20))
-
-		})
 	})
-
-func getNumBadSocketMessages() int {
-	badSocket := regexp.MustCompile(`websocket: bad handshake`)
-	numMessages := 0
-	index, err := pkg.GetOpenSearchSystemIndex(cattleSystemNamespace)
-	if err != nil {
-		pkg.Log(pkg.Error, fmt.Sprintf("Failed to get OpenSearch index: %v", err))
-		return -1
-	}
-
-	template :=
-		`{
-			"size": 1000,
-			"sort": [{"@timestamp": {"order": "desc"}}],
-			"query": {
-				"bool": {
-					"filter" : [
-						{"match_phrase": {"%s": "%s"}},
-						{"range": {"@timestamp": {"gte": "now-%s"}}}
-					]
-				}
-			}
-		}`
-	query := fmt.Sprintf(template, "kubernetes.labels.app.keyword", "rancher", searchTimeWindow)
-	resp, err := pkg.PostElasticsearch(fmt.Sprintf("%s/_search", index), query)
-	if err != nil {
-		pkg.Log(pkg.Error, fmt.Sprintf("Failed to query Elasticsearch: %v", err))
-		return -1
-	}
-	if resp.StatusCode != 200 {
-		pkg.Log(pkg.Error, fmt.Sprintf("Failed to query Elasticsearch: status=%d: body=%s", resp.StatusCode, string(resp.Body)))
-		return -1
-	}
-	var result map[string]interface{}
-	json.Unmarshal(resp.Body, &result)
-
-	hits := pkg.Jq(result, "hits", "hits")
-	if hits == nil {
-		pkg.Log(pkg.Info, "Expected to find hits in log record query results")
-		return -1
-	}
-	pkg.Log(pkg.Info, fmt.Sprintf("Found %d records", len(hits.([]interface{}))))
-	if len(hits.([]interface{})) == 0 {
-		pkg.Log(pkg.Info, "Expected log record query results to contain at least one hit")
-		return -1
-	}
-	for _, h := range hits.([]interface{}) {
-		hit := h.(map[string]interface{})
-		src := hit["_source"].(map[string]interface{})
-		log := src["log"].(string)
-		if badSocket.MatchString(log) {
-			numMessages++
-		}
-	}
-
-	pkg.Log(pkg.Info, fmt.Sprintf("Found %d bad socket messages over the last hour", numMessages))
-	return numMessages
-}
 
 func nsListContains(list []v1.Namespace, target string) bool {
 	for i := range list {
